[main] test: テストが落ちるためstateのテストケースを移動する対応
